% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{app_server}
\alias{app_server}
\title{Application Server}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}
}
\value{

}
\description{
Application Server
}
