% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Risk_data}
\alias{Risk_data}
\title{Covid-age risk factors}
\format{
A data frame with 2072 rows and 4 variables:
\describe{
\item{age}{Age of user.}
\item{name}{Name of the parameter.}
\item{value}{The amount that will be added, or removed, from the users covid-age based on this parameter.}
\item{group}{The group that this parameter belongs to.}
...
}
}
\source{
\url{https://alama.org.uk/covid-19-medical-risk-assessment/}
}
\usage{
Risk_data
}
\description{
A data set of the covid-age risk factors.
}
\details{
A full description of how these were derived can be found
}
\keyword{datasets}
